% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{package}
\name{image.darknet-package}
\alias{image.darknet-package}
\title{Image classification and Object Detection based on darknet}
\description{
Image classification and Object Detection based on darknet
}
\seealso{
\code{\link{image_darknet_classify}}, \code{\link{image_darknet_detect}}, \code{\link{image_darknet_model}}
}
