% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yolo_detect.R
\name{image_darknet_detect}
\alias{image_darknet_detect}
\title{Object detection with YOLO v2 (You only look once)}
\usage{
image_darknet_detect(file, object, threshold = 0.3, hier_threshold = 0.5)
}
\arguments{
\item{file}{character string with the full path to the image file}

\item{object}{an object of class \code{darknet_model} as returned by \code{\link{image_darknet_model}}}

\item{threshold}{numeric, detection threshold}

\item{hier_threshold}{numeric, detection threshold}
}
\value{
invisible(), a file called predictions.png is created in the working directory showing 
the objects found
}
\description{
Object detection with YOLO v2 (You only look once)
}
\examples{
##
## Define the model
##
yolo_tiny_voc <- image_darknet_model(type = 'detect', 
 model = "tiny-yolo-voc.cfg", 
 weights = system.file(package="image.darknet", "models", "tiny-yolo-voc.weights"), 
 labels = system.file(package="image.darknet", "include", "darknet", "data", "voc.names"))
yolo_tiny_voc

##
## Find objects inside the image
##
f <- system.file("include", "darknet", "data", "dog.jpg", package="image.darknet")
x <- image_darknet_detect(file = f, object = yolo_tiny_voc)

\dontrun{
## For other models, see ?image_darknet_model

## trained on COCO
weights <- file.path(system.file(package="image.darknet", "models"), "yolo.weights")
download.file(url = "http://pjreddie.com/media/files/yolo.weights", destfile = weights)
yolo_coco <- image_darknet_model(type = 'detect', 
 model = "yolov3.cfg", 
 weights = system.file(package="image.darknet", "models", "yolov3.weights"), 
 labels = system.file(package="image.darknet", "include", "darknet", "data", "coco.names"))
yolo_coco

f <- system.file("include", "darknet", "data", "dog.jpg", package="image.darknet")
x <- image_darknet_detect(file = f, object = yolo_coco)
}
}
\references{
\url{https://pjreddie.com/publications}
}
\seealso{
\code{\link{image_darknet_model}}
}
