% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/darknet_classify.R
\name{image_darknet_model}
\alias{image_darknet_model}
\title{Specify a model to be used in classification or object detection}
\usage{
image_darknet_model(type = c("classify", "detect"), model, weights, labels,
  resize = TRUE)
}
\arguments{
\item{type}{character string, either 'classify' for classification or 'detect' for object detection}

\item{model}{character string with the full path to the deep learning configuration file}

\item{weights}{character string with the full path to the trained deep learning weights}

\item{labels}{character vector of labels}

\item{resize}{logical indicating to resize the network. Defaults to TRUE. Set to FALSE for the Alexnet and VGG-16 model}
}
\value{
an object of class darknet_model which is a list with these files
}
\description{
Specify a model to be used in classification or object detection.
For classification this consists of
\itemize{
 \item{model: }{the deep learning configuration file which was used to train the model}
 \item{weights: }{the trained weights of the deep learning model}
 \item{labels: }{a character vector of possible labels to classify an image - used during training}
}
}
\examples{
##
## Define the model (structure of the deep learning model + the learned weights + the labels)
##
model <- system.file(package="image.darknet", "include", "darknet", "cfg", "tiny.cfg")
weights <- system.file(package="image.darknet", "models", "tiny.weights")
f <- system.file(package="image.darknet", "include", "darknet", "data", "imagenet.shortnames.list")
labels <- readLines(f)

darknet_tiny <- image_darknet_model(type = 'classify', 
                                    model = model, weights = weights, labels = labels)
darknet_tiny

##
## Other models which are trained already can be downloaded at: 
## https://pjreddie.com/darknet/imagenet
##
\dontrun{
## AlexNet
weights <- file.path(system.file(package="image.darknet", "models"), "alexnet.weights")
download.file(url = "http://pjreddie.com/media/files/alexnet.weights", destfile = weights)
model <- system.file(package="image.darknet", "include", "darknet", "cfg", "alexnet.cfg")

alexnet <- image_darknet_model(type = 'classify', 
                               model = model, weights = weights, labels = labels, resize=FALSE)
alexnet

## Darknet Reference
weights <- file.path(system.file(package="image.darknet", "models"), "darknet.weights")
download.file(url = "http://pjreddie.com/media/files/darknet.weights", destfile = weights)
model <- system.file(package="image.darknet", "include", "darknet", "cfg", "darknet.cfg")

darknetref <- image_darknet_model(type = 'classify', 
                                  model = model, weights = weights, labels = labels)
darknetref

## vgg-16
weights <- file.path(system.file(package="image.darknet", "models"), "vgg-16.weights")
download.file(url = "http://pjreddie.com/media/files/vgg-16.weights", destfile = weights)
model <- system.file(package="image.darknet", "include", "darknet", "cfg", "vgg-16.cfg")

vgg16 <- image_darknet_model(type = 'classify', 
                             model = model, weights = weights, labels = labels, resize=FALSE)
vgg16

## googlenet
weights <- file.path(system.file(package="image.darknet", "models"), "extraction.weights")
download.file(url = "http://pjreddie.com/media/files/extraction.weights", destfile = weights)
model <- system.file(package="image.darknet", "include", "darknet", "cfg", "extraction.cfg")

googlenet <- image_darknet_model(type = 'classify', 
                                 model = model, weights = weights, labels = labels)
googlenet

## darknet19
weights <- file.path(system.file(package="image.darknet", "models"), "darknet19.weights")
download.file(url = "http://pjreddie.com/media/files/darknet19.weights", destfile = weights)
model <- system.file(package="image.darknet", "include", "darknet", "cfg", "darknet19.cfg")

darknet19 <- image_darknet_model(type = 'classify', 
                                 model = model, weights = weights, labels = labels)
darknet19

## Darknet19 448x448
weights <- file.path(system.file(package="image.darknet", "models"), "darknet19_448.weights")
download.file(url = "http://pjreddie.com/media/files/darknet19_448.weights", destfile = weights)
model <- system.file(package="image.darknet", "include", "darknet", "cfg", "darknet19_448.cfg")

darknet_19 <- image_darknet_model(type = 'classify', 
                                  model = model, weights = weights, labels = labels)
darknet_19
}
}
